misc. cleanup / minor fixes
